Refactor: Extract saveSharedReview to utils and remove unused files
- Moved saveSharedReview function from shareMode.ts to dedicated utils file
- Deleted unused shareMode.ts, confirmShare.ts, and shareReview.ts (old dropdown flow)
- Updated imports in confirmCosign.ts, cosignReview.ts, and quoteReview.ts
- Reduces codebase complexity and removes deprecated flow
