+++ src/commands/reviews/buttons/confirmCosign.ts
@@ -1,7 +1,7 @@
 import { MessageComponentInteraction } from 'discord.js'
 import { BotClient } from '../../../Bot'
 import { ReviewType } from '../../../utils/types'
-import { saveSharedReview } from './shareMode'
+import { saveSharedReview } from '../utils/saveSharedReview'
 
 const command = {
   data: { name: 'confirmCosign' },
